BAU: Split form lists generator into separate files
