+++ app/blueprints/assessments/form_lists_generated_translations.py
@@ -1,0 +1,150 @@
+_LIST_TRANSLATIONS_GENERATED = {
+    "Canolfan gymunedol": "Community centre",
+    "Sinema": "Cinema",
+    "Oriel": "Gallery",
+    "Amgueddfa": "Museum",
+    "Lleoliad cerddoriaeth": "Music venue",
+    "Parc": "Park",
+    "Adeilad swyddfa bost": "Post office building",
+    "Tafarn": "Pub",
+    "Siop": "Shop",
+    "Cyfleuster hamdden neu chwaraeon": "sporting",
+    "Theatr": "Theatre",
+    "Arall": "Other",
+    "community-centre": "community-centre",
+    "cinema": "cinema",
+    "gallery": "gallery",
+    "museum": "museum",
+    "music-venue": "music-venue",
+    "park": "park",
+    "post-office": "post-office",
+    "pub": "pub",
+    "shop": "shop",
+    "sporting": "sporting",
+    "theatre": "theatre",
+    "other": "other",
+    "Prynu'r eiddo": "buy-the-asset",
+    "Llogi'r eiddo": "lease-the-asset",
+    "Eisoes yn eiddo i'w sefydliad": "already-owned-by-organisation",
+    "Eisoes wedi'i llogi gan eich sefydliad": "already-leased-by-organisation",
+    "buy-the-asset": "buy-the-asset",
+    "lease-the-asset": "lease-the-asset",
+    "already-owned-by-organisation": "already-owned-by-organisation",
+    "already-leased-by-organisation": "already-leased-by-organisation",
+    "confirm": "confirm",
+    "Ar werth neu wedi'i restru ar gyfer gwaredu": "for-sale-or-listed-for-disposal",
+    "Nid yw'r defnydd yn y dyfodol wedi'i sicrhau": "future-use-not-secured",
+    "Gwerth annyoneddol na thâl fel y mae'r model busnes presennol": "unprofitable-under-current-business-model",
+    "Nid yw'r perchenogaeth gyfredol yn ddichonadwy": "current-ownership-not-tenable",
+    "for-sale-or-listed-for-disposal": "for-sale-or-listed-for-disposal",
+    "future-use-not-secured": "future-use-not-secured",
+    "unprofitable-under-current-business-model": "unprofitable-under-current-business-model",
+    "current-ownership-not-tenable": "current-ownership-not-tenable",
+    "Cau": "Closure",
+    "Werthiant": "Sale",
+    "Camddeall neu wanllyd": "Neglect or dereliction",
+    "Model busnes presennol na ellir ei gynnal": "Unsustainable current business model",
+    "Wedî'i restru ar gyfer gwaredu": "Listed for disposal",
+    "Rhan o Trosglwyddiad Ased Cymunedol": "Part of a Community Asset Transfer",
+    "Closure": "Closure",
+    "Sale": "Sale",
+    "Neglect or dereliction": "Neglect or dereliction",
+    "Unsustainable current business model": "Unsustainable current business model",
+    "Listed for disposal": "Listed for disposal",
+    "Part of a Community Asset Transfer": "Part of a Community Asset Transfer",
+    "Balchder cymunedol a chanfyddiadau o'r ardal leol fel lle i fyw ynddo": "community-pride",
+    "Ymddiriedaeth gymdeithasol, cydlyniant ac ymdeimlad o berthyn": "social-trust",
+    "Cyfranogi mewn bywyd cymunedol, y celfyddydau a diwylliant, neu chwaraeon": "participation",
+    "Canlyniadau economaidd lleol, fel cyflogaeth a chyfleoedd i wirfoddoli, a  chyflogadwyedd a sgiliau": "local-economic", # noqa
+    "Cael effaith gadarnhaol ar iechyd corfforol ac iechyd meddwl, a lleihau unigrwydd ac ynysigrwydd": "delivering-positive", # noqa
+    "community-pride": "community-pride",
+    "social-trust": "social-trust",
+    "participation": "participation",
+    "local-economic": "local-economic",
+    "delivering-positive": "delivering-positive",
+    "Sefydliad corfforedig elusennol (CIO)": "CIO",
+    "Cwmni cydweithredol, fel cymdeithas budd cymunedol": "COOP",
+    "Cwmni buddiannau cymunedol (CIC)": "CIC",
+    "Cwmni cyfyngedig drwy warant": "company-limited",
+    "Sefydliad corfforedig elusennol yr Alban (SCIO)": "SCIO",
+    "Cyngor plwyf, tref neu gymuned": "PTC",
+    "Porthladd ymddiriedolaeth": "trust-port",
+    "CIO": "CIO",
+    "COOP": "COOP",
+    "CIC": "CIC",
+    "CLG": "CLG",
+    "SCIO": "SCIO",
+    "Other": "Other",
+    "PTC": "PTC",
+    "Tŷ'r Cwmnïau": "Companies House",
+    "House": "House",
+    "Yes": "Yes",
+    "No": "No",
+    "Ydy, cyngor tref, plwyf neu gymuned": "Yes, a town, parish or community council",
+    "Ydy, math arall o awdurdod cyhoeddus": "Yes, another type of public authority",
+    "Nac ydy, nid yw'n perthyn i awdurdod cyhoeddus": "No, it does not belong to a public authority",
+    "Cyfleuster chwaraeon neu hamdden": "Sporting or leisure facility",
+    "Penderfyniad i gau neu les yn dod i ben": "Closure or end of lease",
+    "Esgeulustod neu gyflwr adfeiliedig": "Neglect or dereliction",
+    "Model busnes cyfredol anghynaliadwy": "Unsustainable current business model",
+    "Gwerthu": "Sale",
+    "Wedi'i restru i'w waredu": "Listed for disposal",
+    "Rhan o broses trosglwyddo asedau cymunedol": "Part of a community asset transfer",
+    "Ddim yn gwybod pwy yw'r perchennog": "Do not know who owner is",
+    "Yn gwybod pwy yw'r perchennog ond heb gysylltu": "Know who owner is but have not approached",
+    "Trafodaethau ar gam cynnar": "Early stage negotiations",
+    "Trafodaethau ar gam datblygedig": "Advanced stage negotiations",
+    "cadarnhaf": "confirm",
+    "Ysgrifennu cynllun busnes": "Writing a business plan",
+    "Technegol": "Technical",
+    "Rhagolwg ariannol": "Financial forecasting",
+    "Cyfreithiol": "Legal",
+    "Tai": "Housing",
+    "Cymorth llywodraethu": "Governance support",
+    "Rwy'n cytuno y gellir cysylltu â mi os wyf yn gymwys": "Yes, I agree to be contacted if eligible",
+    "Nid wyf am i neb gysylltu â mi": "No, I do not wish to be contacted",
+    "Sefydliad corfforedig elusennol": "Charitable incorporated organisation",
+    "Cydweithrediaeth (er enghraifft, cymdeithas budd cymunedol)": "co-operative",
+    "Cwmni buddiannau cymunedol": "community-interest-company",
+    "Sefydliad corfforedig elusennol yn yr Alban": "scottish-charitable",
+    "Cynghorau plwyf, tref a chymuned": "parish-town-community-councils",
+    "Ddim yn gorfforedig eto": "not-yet-incorporated",
+    "Prynu lesddaliad": "Buy a leasehold",
+    "Adnewyddu lesddaliad": "Renovate a leasehold",
+    "Prynu rhydd-ddaliad": "Buy a freehold",
+    "Adnewyddu rhydd-ddaliad": "Renovate a freehold",
+    "Oes": "Yes",
+    "Nac oes": "No",
+    "Ddim yn siŵr": "Not sure",
+    "Heb ddechrau eto": "Not yet started",
+    "Cam cynnar": "Early stage",
+    "Cam hwyr": "Late stage",
+    "Eisoes wedi'i sicrhau": "Already secured",
+    "Rhoi gwasanaethau proffesiynol": "Donation of professional services",
+    "Rhoi nwyddau": "Donation of goods",
+    "Gostyngiad ar rydd-ddaliad/lesddaliad": "Discount on freehold/leasehold",
+    "Codi arian yn eich cymuned": "Fundraising in your community",
+    "Eich adnoddau ariannol eich hun": "Your own financial resources",
+    "Cyrff cyhoeddus": "Public bodies",
+    "Gweinyddiaethau datganoledig": "Devolved administrations",
+    "Elusennau neu ymddiriedolaethau": "Charities or trusts",
+    "Cyllidwyr y Loteri Genedlaethol": "National lottery funders",
+    "Cyfranddaliadau cymunedol": "Community shares",
+    "Buddsoddwyr cymunedol": "Social Investors",
+    "Benthycwyr eraill": "Other Lenders",
+    "none": "Not sure",
+    "Heb gysylltu ag unrhyw gyllidwyr eto": "Not yet approached any funders",
+    "Wedi cysylltu â rhai cyllidwyr ond heb sicrhau cyllid eto": "Approached some funders but not yet secured",
+    "Wedi cysylltu â'r holl gyllidwyr ond heb sicrhau cyllid eto": "Approached all funders but not yet secured",
+    "Wedi sicrhau rhywfaint o arian cyfatebol": "Secured some match funding",
+    "Wedi sicrhau'r holl arian cyfatebol": "Secured all match funding",
+    "support-local-community": "support-local-community",
+    "improve-local-skills": "improve-local-skills",
+    "net-zero": "net-zero",
+    "regeneration": "regeneration",
+    "fight-climate-change": "fight-climate-change",
+    "one": "one",
+    "two": "two",
+    "yes": "yes",
+    "no": "no",
+}
